[tex2png-hs] Add optional extra packages
o Add _pkgs :: Maybe [Text] to Args to allow optionally including extra
packages (sty files) when generating the tex file.

IMPORTANT: The library doesn't check for existence of the packages, so
make sure packages are correctly installed (or the path to sty files are
indeed correct) when using _pkgs.

o On the command line, the new -P switch is available, and takes a
comma-separated list of packages. e.g. Calling the program with
`-P eventB,unitb`
would populate _pkgs with
Just ["eventB", "unitb"]
which will then be used to generate the usepackage commands:

\usepackage{eventB}
\usepackage{unitb}